Fix tests to workaround supybot issue discovered in #9.
The Owner plugin in 0.83.4.1, for some reason, treats "log" as a command:

    def isCommandMethod(self, name):
        return name == 'log' or \
               self.__parent.isCommandMethod(name)

So it has a "log" command, but it's broken and raises this NoneType is not
callable exception.  The workaround is to set the Git plugin as the primary
plugin for the "log" command, since Owner is broken anyway.